external-deps/envs-manager/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "envs-manager"
description = 'A manager for Python environments and packages managers'
readme = "README.md"
requires-python = ">=3.7"
license = { text = "MIT" }
keywords = []
authors = [
  { name = "Spyder Development Team and envs-manager contributors" },
  { name = "Spyder Development Team and envs-manager contributors", email = "spyder.python@gmail.com" },
]
classifiers = [
  "Development Status :: 2 - Pre-Alpha",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.7",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: Implementation :: CPython",
]
dependencies = [
  "pyyaml",
  "requests"
]
dynamic = ["version"]

[project.urls]
Documentation = "https://github.com/spyder-ide/envs-manager#readme"
Issues = "https://github.com/spyder-ide/envs-manager/issues"
Source = "https://github.com/spyder-ide/envs-manager"

[project.scripts]
envs-manager = "envs_manager.cli:main"

[project.optional-dependencies]
test = [
  "pytest",
  "pytest-cov"
]
pre-commit = [
  "pre-commit"
]

[tool.hatch.version]
path = "envs_manager/__about__.py"

[tool.hatch.build.targets.sdist]
[tool.hatch.build.targets.wheel]

[tool.hatch.envs.default]
dependencies = [
  "pytest",
  "pytest-cov",
]
[tool.hatch.envs.default.scripts]
cov = "pytest --cov-report=term-missing --cov-config=pyproject.toml --cov=envs_manager --cov=tests"
no-cov = "cov --no-cov"

[[tool.hatch.envs.test.matrix]]
python = ["37", "310"]

[tool.coverage.run]
branch = true
parallel = true
omit = [
  "envs_manager/__about__.py",
]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]

[tool.black]
py36 = true
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.github
  | \.hg
  | \.mypy_cache
  | \.pytest_cache
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''
